requirements.txt
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langgraph==0.2.20
tavily-python==0.5.0
tiktoken==0.7.0
pillow==10.4.0
streamlit==1.38.0
watchdog==4.0.2
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.9.0
